pyproject.toml
[project]
name = "ppc-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Jerson Miranda",email = "jerson.miranda@outlook.es"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]
packages = [{include = "ppc_back", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
